Create Expression cheker
A Studnet is given a task to create some mathematical expression for BODMAS. but due to some other work he forgot to do that but in the night he clicked that he have to create some expression for BODMAS due to sleepy mood he did some mistake while putting the brackets. you job is to checK wether the expression give by that boy is Valid or not.

Sample Input: [3+7{52/11(3+5)}] Valid expression as all the brackets are properly open and close [4-6{235(9+6)] Invalid expression as some brackets are not proper.