Create finding supervisors
a tyre factory many workers are working there each worker is assigned with a integer number which help them to find their Supervisor. the project Manager having an Array that consist of a integer numbers help the project Manager to find thier supervisors. Supervisor will be assinged as the first largest integer find on the right side of the array if not found return -1.

i/p: [3,5,2,14,5,3,7,9,4,6,9,4,2,5,3] o/p:=[5,14,14,-1,7,7,9,-1,6,9,-1,5,5,-1,-1]